reusable component created and applied in views
